pyproject.toml
[project]
name = "nutriciolog-qwen"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "aiogram>=3.25.0",
    "gigachat>=0.1.3",
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
    "aiofiles>=24.1.0",
    "certifi>=2024.0.0",
    "httpx>=0.27.0"
]
